+++ pypgx/api/utils.py
@@ -164,9 +164,10 @@ def combine_results(
 
 def compare_genotypes(first, second, verbose=False):
     """
-    Calculate concordance rate between two genotype results.
+    Calculate concordance between two genotype results.
 
-    The method will only use samples that appear in both genotype results.
+    Only samples that appear in both genotype results will be used to
+    calculate concordance for genotype calls as well as CNV calls.
 
     Parameters
     ----------
@@ -177,15 +178,21 @@ def compare_genotypes(first, second, verbose=False):
         Second archive file or object with the semantic type
         SampleTable[Results].
     verbose : bool, default: False
-        If True, print the verbose version of output.
+        If True, print the verbose version of output, including discordant
+        calls.
 
     Examples
     --------
     >>> import pypgx
-    >>> pypgx.compare_genotypes('first-results.zip', 'second-results.zip')
-    Total: 32
-    Compared: 32
-    Concordance: 1.000 (32/32)
+    >>> pypgx.compare_genotypes('results-1.zip', 'results-2.zip')
+    # Genotype
+    Total: 100
+    Compared: 100
+    Concordance: 1.000 (100/100)
+    # CNV
+    Total: 100
+    Compared: 100
+    Concordance: 1.000 (100/100)
     """
     if isinstance(first, str):
         first = sdk.Archive.from_file(first)
@@ -197,22 +204,24 @@ def compare_genotypes(first, second, verbose=False):
 
     second.check_type('SampleTable[Results]')
 
-    df = pd.concat([first.data.Genotype, second.data.Genotype], axis=1)
+    def show_comparison(col):
+        df = pd.concat([first.data[col], second.data[col]], axis=1)
+        print(f'# {col}')
+        print(f'Total: {df.shape[0]}')
+        df.columns = ['First', 'Second']
+        df = df.dropna()
+        print(f'Compared: {df.shape[0]}')
+        df['Concordant'] = df.First == df.Second
+        print(f'Concordance: {sum(df.Concordant)/len(df.Concordant):.3f} ({sum(df.Concordant)}/{len(df.Concordant)})')
+        if verbose:
+            print('Discordant:')
+            if df.Concordant.all():
+                print('None')
+            else:
+                print(df[~df.Concordant])
 
-    print(f'Total: {df.shape[0]}')
-
-    df.columns = ['First', 'Second']
-    df = df.dropna()
-
-    print(f'Compared: {df.shape[0]}')
-
-    df['Concordant'] = df.First == df.Second
-
-    print(f'Concordance: {sum(df.Concordant)/len(df.Concordant):.3f} ({sum(df.Concordant)}/{len(df.Concordant)})')
-
-    if verbose:
-        print('Discordant genotypes:')
-        print(df[~df.Concordant])
+    for col in ['Genotype', 'CNV']:
+        show_comparison(col)
 
 def compute_control_statistics(
     bam=None, fn=None, gene=None, region=None, assembly='GRCh37', bed=None